+++ src/main/java/com/github/smallinger/coppergolemlegacy/entity/CopperGolemState.java
@@ -15,7 +15,8 @@ public enum CopperGolemState implements StringRepresentable {
     GETTING_ITEM("getting_item", 1),
     GETTING_NO_ITEM("getting_no_item", 2),
     DROPPING_ITEM("dropping_item", 3),
-    DROPPING_NO_ITEM("dropping_no_item", 4);
+    DROPPING_NO_ITEM("dropping_no_item", 4),
+    PRESSING_BUTTON("pressing_button", 5);
 
     public static final Codec<CopperGolemState> CODEC = StringRepresentable.fromEnum(CopperGolemState::values);
     private static final IntFunction<CopperGolemState> BY_ID = ByIdMap.continuous(CopperGolemState::id, values(), ByIdMap.OutOfBoundsStrategy.ZERO);